Bug fix for <max on cells> 9-12
The search ended early if <max on cells> was in the range 9 to 12 bits.
This was related to subset division being confused - it uses the 9 first
defined on-cells to define the current subset.
Searches including bit counts 9-12 but with a higher <max on cells> than
that should not have been affected.